requirements.txt
transformers
python-docx

